requirements.txt
flask==2.0.2
sqlalchemy==1.4.27
marshmallow-sqlalchemy==0.26.1
Flask-Restless-NG==2.2.9
flask-restx==0.5.1
flask-login==0.5.0
fido2==0.9.3
flask-marshmallow==0.14.0
flask-sqlalchemy==2.5.1